refactor: centralize issue run contracts
