+++ src/cli/commands/run-once/implementation-agent.ts
@@ -14,15 +14,15 @@ import {
 } from "./prompts.ts";
 import type { IssueWorktreeResult } from "./git.ts";
 import type { CommandRunner } from "../../../command/types.ts";
-import type { IssueSummary } from "../../../issue/types.ts";
 import type {
   AgentIssueBlockedResult,
   AgentIssueDevelopmentEnvironmentNotReadyResult,
   AgentIssueMergedResult,
   AgentIssuePiResult,
   AgentIssuePrCreatedResult,
-  AgentIssueRunState,
-} from "./types.ts";
+} from "../../../issue-run/types.ts";
+import type { IssueSummary } from "../../../issue/types.ts";
+import type { AgentIssueRunState } from "./types.ts";
 import type { AgentIssueProgressEvent, ProgressReporter } from "./progress.ts";
 import type { AgentIssueConfig } from "./types.ts";
