+++ src/cli/commands/run-once/planning-phase-coordinator.ts
@@ -1,10 +1,10 @@
 import { planningPhasePlan } from "../../../workflow/planning-pull-requests.ts";
 import type { PlanningStateV1 } from "../../../workflow/planning-state-types.ts";
-import type { IssueSummary } from "../../../issue/types.ts";
 import type {
   AgentIssueBlockedResult,
   AgentIssuePrCreatedResult,
-} from "./types.ts";
+} from "../../../issue-run/types.ts";
+import type { IssueSummary } from "../../../issue/types.ts";
 import type { PlanningPhaseRunnerOutcome } from "./planning-phase-runner.ts";
 
 export type PlanningCoordinatorOutcome =
