+++ src/cli/commands/run-once/types.ts
@@ -7,13 +7,19 @@ import type { PatchmillSkillsConfig } from "../../../workflow/skills.ts";
 import type { RunCostReport } from "./run-cost.ts";
 import type { CommandRunner } from "../../../command/types.ts";
 import type { IssueSummary } from "../../../issue/types.ts";
+import type {
+  AgentIssueBlockedResult,
+  AgentIssueBlockerQuestion,
+  AgentIssueMergedResult,
+  AgentIssuePrCreatedResult,
+  AgentIssueVisualEvidence,
+} from "../../../issue-run/types.ts";
 
 export type {
   CommandResult,
   CommandRunOptions,
   CommandRunner,
 } from "../../../command/types.ts";
-export type { HumanDecisionQuestion } from "../triage/types.ts";
 export type {
   IssueCommentSummary,
   IssueSummary,
@@ -21,6 +27,22 @@ export type {
   LabelDefinition,
 } from "../../../issue/types.ts";
 export type { AgentIssueProgressEvent, ProgressReporter } from "./progress.ts";
+export type {
+  AgentIssueBlockedResult,
+  AgentIssueBlockerQuestion,
+  AgentIssueDevelopmentEnvironmentHandoff,
+  AgentIssueDevelopmentEnvironmentNotReadyResult,
+  AgentIssueDevelopmentEnvironmentReadyResult,
+  AgentIssueDevelopmentEnvironmentResult,
+  AgentIssueImplementationResumeContext,
+  AgentIssueMergedResult,
+  AgentIssuePiResult,
+  AgentIssuePlanCreatedResult,
+  AgentIssuePrCreatedResult,
+  AgentIssueSpecCreatedResult,
+  AgentIssueVisualEvidence,
+} from "../../../issue-run/types.ts";
+export type { HumanDecisionQuestion } from "../../../workflow/decisions.ts";
 
 export type AgentIssueConfig = {
   repoRoot: string;
@@ -190,85 +212,11 @@ export type AgentIssueRunStateUpdate = {
   clearBlockerQuestions?: boolean | undefined;
 };
 
-export type AgentIssueImplementationResumeContext = {
-  resumed: boolean;
-  worktreeCreated: boolean;
-  existingCommits: string[];
-  priorBlockerReason?: string | undefined;
-  priorBlockerQuestions?: AgentIssueBlockerQuestion[] | undefined;
-  priorValidation?: string[] | undefined;
-};
-
-export type AgentIssueBlockerQuestion =
-  | string
-  | import("../triage/types.ts").HumanDecisionQuestion;
-
-export type AgentIssueBlockedResult = {
-  status: "blocked";
-  reason: string;
-  questions: AgentIssueBlockerQuestion[];
-  commits: string[];
-  validation: string[];
-};
-
-export type AgentIssueSpecCreatedResult = {
-  status: "spec-created";
-  specPath: string;
-  commit?: string | undefined;
-};
-
-export type AgentIssuePlanCreatedResult = {
-  status: "plan-created";
-  planPath: string;
-  commit?: string | undefined;
-};
-
 export type AgentIssueApprovalRequiredResult = {
   status: "approval-required";
   issue: IssueSummary;
   approvalKind: "spec" | "plan";
   missingLabel: string;
-};
-
-export type AgentIssueDevelopmentEnvironmentReadyResult = {
-  status: "ready";
-  summary: string;
-  evidence: string[];
-  environment?: Record<string, string> | undefined;
-};
-
-export type AgentIssueDevelopmentEnvironmentNotReadyResult = {
-  status: "not-ready";
-  reason: string;
-  evidence: string[];
-  remediation: string[];
-};
-
-export type AgentIssueDevelopmentEnvironmentResult =
-  | AgentIssueDevelopmentEnvironmentReadyResult
-  | AgentIssueDevelopmentEnvironmentNotReadyResult;
-
-export type AgentIssueDevelopmentEnvironmentHandoff =
-  AgentIssueDevelopmentEnvironmentReadyResult & {
-    completedAt: string;
-  };
-
-export type AgentIssueVisualEvidence = {
-  screenshotPath: string;
-  caption?: string | undefined;
-  referencePaths?: string[] | undefined;
-  url?: string | undefined;
-};
-
-export type AgentIssuePrCreatedResult = {
-  status: "pr-created";
-  prUrl: string;
-  branch: string;
-  commits: string[];
-  validation: string[];
-  reviewSummary?: string | undefined;
-  landingDecision?: string | undefined;
-  visualEvidence?: AgentIssueVisualEvidence[] | undefined;
 };
 
 export type AgentIssueReviewPendingResult = {
@@ -287,23 +235,6 @@ export type AgentIssueStoppedResult = {
   branch?: string | undefined;
   worktreePath?: string | undefined;
 };
-
-export type AgentIssueMergedResult = {
-  status: "merged";
-  branch: string;
-  mergeCommit: string;
-  commits: string[];
-  validation: string[];
-  reviewSummary?: string | undefined;
-  landingDecision?: string | undefined;
-};
-
-export type AgentIssuePiResult =
-  | AgentIssueBlockedResult
-  | AgentIssueSpecCreatedResult
-  | AgentIssuePlanCreatedResult
-  | AgentIssuePrCreatedResult
-  | AgentIssueMergedResult;
 
 type AgentIssuePipelineResultLog = {
   logPath?: string | undefined;
