+++ src/workflow/decisions.ts
@@ -1,0 +1,4 @@
+export type HumanDecisionQuestion = {
+  question: string;
+  recommendedAnswer: string;
+};
